wvcsv.py: clean up the API for quoting and iterating.
wvcsv.Reader() now yields a series of tuples, so you don't need to call
wvcsv.splitline() manually.  We can thus hide a bunch of functions back in
_wvcsv so people know not to depend on them.

Add wvcsv.quotel(), which quotes an entire list of objects and returns a
line (including terminating \r\n) suitable for writing to a .csv file.